requirements.txt
# Core Dependencies
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.0
groq>=0.4.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
python-dotenv>=1.0.0

# Web Search & Scraping
google-search-results>=2.4.2
duckduckgo-search>=4.1.1
playwright>=1.41.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.1.0
aiohttp>=3.9.0

# Document Generation
reportlab>=4.0.9
python-docx>=1.1.0
markdown>=3.5.2
pypdf>=4.0.1

# Data & Vector Storage
faiss-cpu>=1.7.4
numpy>=1.26.0
pandas>=2.2.0

# Agent Orchestration
langgraph>=0.0.20
pydantic>=2.6.0

# Utilities
colorama>=0.4.6
rich>=13.7.0
tqdm>=4.66.0
